requirements.txt
streamlit
pandas
plotly
numpy
numpy-financial
openrouteservice
python-dotenv
beautifulsoup4
